updates
Archive. Let’s let it live forever for free without any maintenance. Is it possible to make the following changes? 1. Make the navigation above the fold on the home page, 2) “their own pieces” of the story (add an s to pieces) on the About page, 3) contact form to go to ---), 4) remove the dash between 2926 and 16th Street in the address